Move sandbox utils to match habitat-lab main.
