+++ examples/siro_sandbox/app_states/app_state_tutorial.py
@@ -7,8 +7,8 @@
 from app_states.app_state_abc import AppState
 from hitl_tutorial import Tutorial, generate_tutorial
 
-from habitat.gui.gui_input import GuiInput
-from habitat.gui.text_drawer import TextOnScreenAlignment
+from utils.gui.gui_input import GuiInput
+from utils.gui.text_drawer import TextOnScreenAlignment
 
 
 class AppStateTutorial(AppState):
