+++ examples/siro_sandbox/camera_helper.py
@@ -9,7 +9,7 @@ from typing import Tuple
 import magnum as mn
 import numpy as np
 
-from habitat.gui.gui_input import GuiInput
+from utils.gui.gui_input import GuiInput
 
 
 class CameraHelper:
