+++ examples/siro_sandbox/gui_navigation_helper.py
@@ -5,7 +5,7 @@
 # LICENSE file in the root directory of this source tree.
 
 import magnum as mn
-from hablab_utils import get_agent_art_obj_transform
+from utils.hablab_utils import get_agent_art_obj_transform
 
 import habitat_sim
 from habitat.datasets.rearrange.navmesh_utils import get_largest_island_index
