+++ examples/siro_sandbox/server/remote_gui_input.py
@@ -7,7 +7,7 @@
 import magnum as mn
 from server.average_rate_tracker import AverageRateTracker
 
-from habitat.gui.gui_input import GuiInput
+from utils.gui.gui_input import GuiInput
 
 
 class RemoteGuiInput:
