+++ examples/siro_sandbox/test_episode_save_files.py
@@ -7,7 +7,7 @@
 import argparse
 
 import numpy as np
-from serialize_utils import load_json_gzip, load_pickle_gzip
+from utils.serialize_utils import load_json_gzip, load_pickle_gzip
 
 
 def get_comp_value(a):
