+++ examples/siro_sandbox/utils/gui/gui_application.py
@@ -11,8 +11,8 @@ import time
 import magnum as mn
 from magnum.platform.glfw import Application
 
-from habitat.gui.gui_input import GuiInput
-from habitat.gui.text_drawer import TextOnScreenAlignment
+from utils.gui.gui_input import GuiInput
+from utils.gui.text_drawer import TextOnScreenAlignment
 
 
 class GuiAppDriver:
