+++ examples/siro_sandbox/utils/gui/image_framebuffer_drawer.py
