+++ examples/siro_sandbox/utils/gui/text_drawer.py
